Fixed up most of the tests, about ready to PR
